Rename boards/ folder to components/
