+++ packages/gemini-kit/src/components/gemini-generator.ts
