+++ packages/gemini-kit/src/components/nano-generator.ts
